requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.13.0
python-dotenv>=1.0.0
crewai>=0.1.0
chromadb>=0.4.0
python-dateutil>=2.8.2
streamlit>=1.32.0
plotly>=5.18.0
kafka-python>=2.0.2
confluent-kafka>=2.3.0
aiokafka>=0.10.0
faust-streaming>=0.10.0
scikit-learn>=1.4.0 